alternative os: create build and deploy jobs for fedora and debian

+++ docker/scripts/install_docker_dependencies.sh
@@ -20,7 +20,7 @@ install_docker_dependencies_debian_ubuntu() {
     local -a pkgs=()
 
     local -a build_pkgs=(git ca-certificates)
-    local -a run_pkgs=(curl ntpdate tini)
+    local -a run_pkgs=(curl tini)
 
     case "$mode" in
         build)
@@ -36,6 +36,13 @@ install_docker_dependencies_debian_ubuntu() {
     esac
 
     DEBIAN_FRONTEND=noninteractive apt-get update
+    if [[ "$mode" == "run" ]]; then
+        if apt-cache policy ntpdate | awk '$1 == "Candidate:" && $2 != "(none)" { found = 1 } END { exit !found }'; then
+            pkgs+=(ntpdate)
+        else
+            pkgs+=(ntpsec-ntpdate)
+        fi
+    fi
     DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends "${pkgs[@]}"
     apt-get clean && rm -rf /var/lib/apt/lists/*
 }
@@ -44,7 +51,7 @@ install_docker_dependencies_arch() {
     local mode="${1:?}"
     local -a pkgs=()
 
-    local -a build_pkgs=(git ca-certificates)
+    local -a build_pkgs=(git ca-certificates base-devel which)
     local -a run_pkgs=(curl ntp)
 
     case "$mode" in
@@ -68,8 +75,8 @@ install_docker_dependencies_fedora() {
     local mode="${1:?}"
     local -a pkgs=()
 
-    local -a build_pkgs=(git ca-certificates)
-    local -a run_pkgs=(curl ntp)
+    local -a build_pkgs=(git ca-certificates make gcc gcc-c++ pkgconf-pkg-config which)
+    local -a run_pkgs=(curl chrony)
 
     case "$mode" in
         build)
@@ -92,7 +99,7 @@ install_docker_dependencies_rhel() {
     local mode="${1:?}"
     local -a pkgs=()
 
-    local -a build_pkgs=(git ca-certificates)
+    local -a build_pkgs=(git ca-certificates make gcc gcc-c++ pkgconf-pkg-config which)
     local -a run_pkgs=(curl ntp)
 
     case "$mode" in
